Don't crash if info->deleteProc == NULL

+++ generic/itclEnsemble.c
@@ -87,10 +87,8 @@ typedef struct EnsembleParser {
     Ensemble* ensData;            /* add parts to this ensemble */
 } EnsembleParser;
 
-static int EnsembleSubCmd(void *clientData, Tcl_Interp *interp,
-        int objc, Tcl_Obj *const objv[]);
-static int EnsembleUnknownCmd(void *dummy, Tcl_Interp *interp,
-    int objc, Tcl_Obj *const objv[]);
+static Tcl_ObjCmdProc EnsembleSubCmd;
+static Tcl_ObjCmdProc EnsembleUnknownCmd;
 
 /*
  *  Forward declarations for the procedures used in this file.
@@ -360,7 +358,7 @@ ensPartFail:
 #if TCL_MAJOR_VERSION > 8
 
 typedef struct {
-	Tcl_ObjCmdProc2 *objProc;
+    Tcl_ObjCmdProc2 *objProc;
     void *clientData;
     Tcl_CmdDeleteProc *deleteProc;
 } ensInfo;
@@ -369,18 +367,20 @@ static int ensCmdProc(
     void *clientData,
     Tcl_Interp *interp,
     int argc,
-	Tcl_Obj *const *objv)
+    Tcl_Obj *const *objv)
 {
-	ensInfo *info = (ensInfo *)clientData;
+    ensInfo *info = (ensInfo *)clientData;
     return info->objProc(info->clientData, interp, argc, objv);
 }
 
 static void ens2DeleteProc(
     void *clientData)
 {
-	ensInfo *info = (ensInfo *)clientData;
+    ensInfo *info = (ensInfo *)clientData;
+    if (info->deleteProc != NULL) {
 	info->deleteProc(info->clientData);
-	ckfree(info);
+    }
+    ckfree(info);
 }
 
 int
@@ -393,10 +393,10 @@ Itcl_AddEnsemblePart2(
     void *clientData,              /* client data associated with part */
     Tcl_CmdDeleteProc *deleteProc) /* procedure used to destroy client data */
 {
-	ensInfo *info = (ensInfo *)ckalloc(sizeof(ensInfo));
-	info->objProc = objProc;
-	info->clientData = clientData;
-	info->deleteProc = deleteProc;
+    ensInfo *info = (ensInfo *)ckalloc(sizeof(ensInfo));
+    info->objProc = objProc;
+    info->clientData = clientData;
+    info->deleteProc = deleteProc;
     return Itcl_AddEnsemblePart(interp, ensName, partName, usageInfo, ensCmdProc, info, ens2DeleteProc);
 }
 #endif /* TCL_MAJOR_VERSION */
@@ -2119,15 +2119,15 @@ CallInvokeEnsembleMethod2(
     int result)
 {
     EnsemblePart *ensPart = (EnsemblePart *)data[0];
-    int objc = PTR2INT(data[1]);
-    Tcl_Obj *const*objv = (Tcl_Obj *const*)data[2];
+    Tcl_Size objc = PTR2INT(data[1]);
+    Tcl_Obj *const *objv = (Tcl_Obj *const*)data[2];
     result = (*ensPart->objProc)(ensPart->clientData, interp, objc, objv);
     return result;
 }
 
 static int
 EnsembleSubCmd(
-    void *clientData,      /* ensPart struct pointer */
+    void *clientData,           /* ensPart struct pointer */
     Tcl_Interp *interp,         /* Current interpreter. */
     int objc,                   /* Number of arguments. */
     Tcl_Obj *const objv[])      /* Argument objects. */
